requirements.txt
Flask==2.3.3
pymongo==4.6.1
werkzeug==2.3.7
python-dotenv==1.0.1
azure-ai-formrecognizer==3.3.0
gunicorn==21.2.0
pyngrok==7.0.0
Flask-Login==0.6.2